Add thermistor computeTemp tests
