+++ platformio.ini
@@ -9,5 +9,9 @@ lib_deps =
     olikraus/U8g2@^2.33.2
     https://github.com/GyverLibs/GyverStepper.git
     https://github.com/gyverlibs/GyverEncoder.git
-	https://github.com/gyverlibs/GyverPID.git
-	teemuatlut/TMCStepper@^0.7.3
+        https://github.com/gyverlibs/GyverPID.git
+        teemuatlut/TMCStepper@^0.7.3
+
+[env:native]
+platform = native
+build_src_filter = +<thermistor.cpp>
